Entity classes inlined into Test Ignore phpcs validation as it doesn't support property(set) yet.
